Desafio 13,14,15 feito

+++ Desafio11.py
@@ -2,8 +2,8 @@
 # calcule a sua area e a quantidade de tinta necessaria para pintá-lo, 
 # sabendo que cada litro de tinta pinta uma arra de 2m²
 
-altura = int(input('Qual é a altura da parede? '))
-largura = int(input('Qual é a largura da parede? '))
+altura = float(input('Qual é a altura da parede? '))
+largura = float(input('Qual é a largura da parede? '))
 area = altura * largura
 litroTinta = 2
 quantidade = area / litroTinta
